bug correction - attempt/trace mixup
fixed a big problem where urls was asking for a trace_id and received an attempt_id instead. worked for a moment unnoticed because the ids are all numbers starting at 1 and the id technically existed.
Now the TraceImage doesn't get linked to a Trace right away and this is instead done at the Trace creation.

With this method we can also periodically remove all images that are not linked to a trace, which means they were uploaded but not sent to the model, and are not useful to keep.