modified cropper
fidgeted around a little with the cropper to test further and the problems i discussed with Renaud came from the "dragMode". it was set to crop, so each time the user tries to grab the corner of the crop area and doesn't touch the corner, it will reset. A bit complicated to explain on text, the important part is I found it annoying to use and changed it. It wil be very easy to reswitch if this is not the test users opinion anyway.

Also, the page did not scale properly on my phone, i couldn't see the whole buttons so i added a little scaler.