little modifications before the interview
## 25.11 INTERVIEWS COMMIT ##
these changes were made yesterday before the interviews in a few tests to ensure said interviews go well.
The code is temporary (and not very clean), I mainly added a failsafe in case the format wasn't respected. If said case was to happen, the whole answer would be dumped in the displayed response instead of crashing so the tests could continue.

There is also two prompt modifications that will be remade cleaner too, this commit is only here to have a "state of the code for the interviews" commit.