attempts import/export
added two django management commands:
- "export_attempt" to export one attempt from the db to a local json file structured as in /template/export/attempt.json.
The user must input the attempt id. It is intended to be used side by side with a db visualizer (like beekeeper) to get said id, an upgrade could be to ask the user which student/exercise he wants to export but for the moment this version should suffice.

- "import_attempt" to import an attempt from a json file structured as in /template/export/attempt.json to the db.
The user must input a file path to a valid json. The import happens as a transaction: if anything goes wrong, nothing will be imported in the db (instead of leaving the part of the import that worked)