[archiving] first attempt at highlighted image in interface
- adds a field "image_source" in TraceImage to differentiate the pictures that were uploaded by the user and the ones that were modified by the AI, so the AI modified images can be stored as TraceImage in the first place.
- adds the AI modified image to the chat under the LLM response in both the computer and phone version

This display method and image storage was discussed with Renaud and some big adjustments are to be made, this commit serves as an archive "checkpoint" and is not to be pushed on main. If it needs to be tested, it also requires to do a django migration beforehand, since a field was added to TraceImage.